test cases updated and response rectification

+++ payments/domain/tests/test_dapp_manager.py
@@ -12,17 +12,17 @@ class TestDappMananger(unittest.TestCase):
     dapp_order_manager = OrderManager()
 
     def test_create_order(self):
-        response = self.dapp_order_manager.create_order(100, "USD", {"item": "item", "quantity": 2}, "user")
-        assert "body" in response
-        assert "headers" in response
-        assert "statusCode" in response
-        assert "order_id" in response["body"]
-        assert "item_details" in response["body"]
-        assert "amount" in response["body"]
-        assert "currency" in response["body"]
-        order_id = response["body"]["order_id"]
-        print(response)
-        self.repo.session.query(Order.id == order_id).delete()
+        status, response = self.dapp_order_manager.create_order(100, "USD", {"item": "item", "quantity": 2}, "user")
+        if status:
+            assert "order_id" in response
+            assert "item_details" in response
+            if "amount" in response:
+                assert "amount" in response["amount"]
+                assert "currency" in response["amount"]
+            else:
+                assert False
+            order_id = response["order_id"]
+            self.repo.session.query(Order.id == order_id).delete()
 
     def test_initiate_order(self):
         self.repo.add_item(
@@ -40,27 +40,25 @@ class TestDappMananger(unittest.TestCase):
                 username="user",
             )
         )
-        response = self.dapp_order_manager.initiate_payment_against_order("order_test_123", 100, "USD", "paypal")
-        assert "body" in response
-        assert "headers" in response
-        assert "statusCode" in response
-        assert "amount" in response["body"]
-        assert "currency" in response["body"]
-        assert "payment_id" in response["body"]
-        assert "payment" in response["body"]
-        assert "order_id" in response["body"]
+        status, response = self.dapp_order_manager.initiate_payment_against_order("order_test_123", 100, "USD", "paypal")
+        if status:
+            if "amount" in response:
+                assert "amount" in response["amount"]
+                assert "currency" in response["amount"]
+            else:
+                assert False
+            assert "payment_id" in response
+            assert "payment" in response
+            assert "order_id" in response
 
     def test_execute_order(self):
         pass
 
     def test_order_details(self):
         username = "123"
-        order_details = self.dapp_order_manager.get_order_details_for_user(username)
-        assert "body" in order_details
-        assert "headers" in order_details
-        assert "statusCode" in order_details
-        if order_details["statusCode"] == "success":
-            assert "orders" in order_details["body"]
+        status, order_details = self.dapp_order_manager.get_order_details_for_user(username)
+        if status:
+            assert "orders" in order_details
 
 
 if __name__ == "__main__":
